+++ packer/debian-7.4.0-amd64/scripts/docker.sh
@@ -1,33 +1,19 @@
 #!/bin/bash -eux
 
+# enable memory and swap cgroup
+perl -p -i -e 's/GRUB_CMDLINE_LINUX=""/GRUB_CMDLINE_LINUX="cgroup_enable=memory swapaccount=1"/g'  /etc/default/grub
+/usr/sbin/update-grub
+
+# add docker group and add vagrant to it
+groupadd docker
+gpasswd -a vagrant docker
+
 apt-get -y update
-apt-get -y install golang aufs-tools libdevmapper-dev
+apt-get -y install curl
+apt-get -y install linux-image-extra-`uname -r`
 
-# get rid of interactive prompt
-echo 'lxc lxc/directory string /var/lib/lxc' > /tmp/lxc_debconf
-sudo debconf-set-selections -v /tmp/lxc_debconf
-apt-get -y install lxc
-rm -f /tmp/lxc_debconf
+echo deb http://get.docker.io/ubuntu docker main | sudo tee /etc/apt/sources.list.d/docker.list
+sudo apt-key adv --keyserver keyserver.ubuntu.com --recv-keys 36A1D7869245C8950F966E92D8576A8BA88D21E9
 
-export GOPATH=~/usr/lib/go
-export PATH=$GOPATH/bin:$PATH
-
-mkdir -p "$GOPATH"
-
-#GO GET docker...
-go get -v github.com/dotcloud/docker
-#rm -rf $GOPATH/src/github.com/dotcloud/docker/vendor/src/code.google.com/p/go.net/ipv6
-
-#GO INSTALL...
-go install -v github.com/dotcloud/docker/
-
-#Mounting...
-echo 'none /sys/fs/cgroup cgroup defaults 0 0' | sudo tee -a /etc/fstab
-sudo mount /sys/fs/cgroup
-
-#Properly installing docker from binaries...
-wget https://get.docker.io/builds/Linux/x86_64/docker-latest -O /usr/local/bin/docker
-chmod +x /usr/local/bin/docker
-
-#RUN...
-#docker run -i -t ubuntu /bin/bash
+apt-get -y update
+apt-get -y install lxc-docker